pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compression-safeguards"
version = "1.0.0"
description = "Fearless lossy compression using safeguards"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy~=2.0",
    "numpy-quaddtype~=0.2.0; sys_platform != 'emscripten'",
    "semver~=3.0",
    "sly~=0.5.0",
    "typing-extensions~=4.6",
]

[dependency-groups]
bench = [
    "earthkit-data~=0.14.0",
    "earthkit-plots~=0.3.1",
    "matplotlib~=3.10.0",
    "mpl-ascii~=0.10.0",
    "netcdf4~=1.7.2",
    "numcodecs-wasm~=0.2.1",
    "numcodecs-wasm-sz3~=0.8.1",
    "numcodecs-wasm-zfp~=0.6.0",
    "numcodecs-wasm-zstd~=0.4.0",
    "tqdm~=4.0",
    "xarray~=2025.4",
]
dev = ["ruff~=0.9"]
examples = [
    "cdsapi~=0.7.6",
    "earthkit-data~=0.14.0",
    "earthkit-plots~=0.3.1",
    "humanize~=4.14.0",
    "jupyterlab",
    "matplotlib~=3.10.0",
    "netcdf4~=1.7.2",
    "numcodecs-wasm~=0.2.1",
    "numcodecs-wasm-sperr~=0.2.0",
    "numcodecs-wasm-sz3~=0.8.1",
    "numcodecs-wasm-zfp~=0.6.0",
    "numcodecs-zero~=0.1.0",
    # FIXME: https://github.com/dask/dask/issues/12072
    "pyarrow~=21.0.0",
    "scikit-image~=0.25.0",
    "xarray~=2025.1",
]
fuzz = ["atheris~=2.3", "setuptools~=80.0", "timeoutcontext~=1.2"]
test = [
    "dask~=2025.4",
    "numcodecs-combinators~=0.2.10",
    "pytest~=8.4",
    "xarray~=2025.4",
    "zarr~=3.0; python_version >= '3.11'",
    "zarr-any-numcodecs~=0.1.2; python_version >= '3.11'",
]
doctest = ["griffe2md~=1.1", "pytest~=8.3", "pytest-codeblocks~=0.17.0"]
typing = ["mypy~=1.18"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"compression_safeguards" = ["py.typed"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "NPY", "T20", "SLOT", "TC", "PLC", "UP", "FURB"]
ignore = ["PLC0105"]

[tool.ruff.lint.per-file-ignores]
"bench/*" = ["T20"]
"examples/*" = ["T20", "PLC"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
xfail_strict = true

[[tool.mypy.overrides]]
module = ["numpy_quaddtype.*", "sly.*"]
follow_untyped_imports = true

[tool.hatch.build.targets.sdist]
include = ["/src", "/CITATION.cff", "/LICENSE", "/pyproject.toml", "/README.md"]

[tool.uv.workspace]
members = ["packages/numcodecs-safeguards", "packages/xarray-safeguards"]
